+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/StatisticsApiHttpService.scala
@@ -9,20 +9,20 @@ import pl.touk.nussknacker.ui.api.description.StatisticsApiEndpoints.Dtos.{
   StatisticUrlResponseDto
 }
 import pl.touk.nussknacker.ui.db.timeseries.{FEStatisticsRepository, WriteFEStatisticsRepository}
-import pl.touk.nussknacker.ui.security.api.AuthenticationResources
+import pl.touk.nussknacker.ui.security.api.AuthManager
 import pl.touk.nussknacker.ui.statistics.UsageStatisticsReportsSettingsService
 
 import scala.concurrent.{ExecutionContext, Future}
 
 class StatisticsApiHttpService(
-    authenticator: AuthenticationResources,
+    authManager: AuthManager,
     usageStatisticsReportsSettingsService: UsageStatisticsReportsSettingsService,
     repository: FEStatisticsRepository[Future]
 )(implicit ec: ExecutionContext)
-    extends BaseHttpService(authenticator)
+    extends BaseHttpService(authManager)
     with LazyLogging {
 
-  private val endpoints                = new StatisticsApiEndpoints(authenticator.authenticationMethod())
+  private val endpoints                = new StatisticsApiEndpoints(authManager.authenticationEndpointInput())
   private val ignoringErrorsRepository = new IgnoringErrorsStatisticsRepository(repository)
 
   expose {